[master] 3/4 Annuity payment
Description

Let's compute all the parameters of the loan. There are at least two kinds of loan: those with annuity payment and those with differentiated payment. In this stage, you are going to calculate only the former. Annuity type of payment consists in paying a fixed sum of money at specified intervals, for example, each month or each year. The annuity payment amount is precisely this fixed sum of money that you need to pay at regular intervals.

Objectives

In this stage, you should change the behavior of the calculator:

1. First, you should parse the provided parameters to define, which ones are known and which one is missing using the flag package.
2. Compute the missing value using the formulas mentioned above. The calculator should be able to calculate the number of monthly payments, the monthly payment amount, and the loan principal.
3. Finally, output the value users wanted to compute.

The final version of your program is supposed to work from the command line and parse the following arguments:

    --payment is the payment amount. It can be calculated using the provided principal, interest, and number of months
    --principal You can get its value if you know the interest, annuity payment, and number of months.
    --periods denotes the number of months needed to repay the loan. It's calculated based on the interest, annuity payment, and principal.
    --interest is specified without a percent sign. Note that it can accept a floating-point value. Our loan calculator can't calculate the interest, so it must always be provided.