[master] 4/4 Differentiate payment
Description

Finally, let's add to our calculator the capacity to compute differentiated payments. We’ll do this for the type of repayment where the loan principal is reduced by a constant amount each month. The rest of the monthly payment goes toward interest repayment and it is gradually reduced over the term of the loan.

Objectives

In this stage, you are going to implement the following features:

1. Calculation of differentiated payments. To do this, the user can run the program specifying interest, number of monthly payments, and loan principal.
2. Ability to calculate the same values as in the previous stage for annuity payment (principal, number of monthly payments, and monthly payment amount).
3. Handling of invalid parameters. It's a good idea to show an error message if the user enters invalid parameters.
4. The only thing left is to compute the overpayment: the amount of interest paid over the whole term of the loan.